rag/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
fastar==0.8.0
google-auth==2.49.1
google-genai==1.68.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jinja2==3.1.6
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
numpy==2.4.3
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pygments==2.19.2
python-dotenv==1.2.2
python-multipart==0.0.22
pyyaml==6.0.3
requests==2.32.5
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.55.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.52.1
tenacity==9.1.4
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
Pillow>=10.0.0
ollama
pandas
scikit-learn
